\name{ChIPKernels-package}
\alias{ChIPKernels-package}
\alias{ChIPKernels}
\docType{package}
\title{ String kernels for ChIP-Seq
}
\description{
Collection of methods for building different string kernels used for
    ChIP-Seq.}
\details{
\tabular{ll}{
Package: \tab ChIPKernels\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-21\cr
License: \tab What license is it under?\cr
}
}
\author{
Manu Setty

Maintainer: Manu Setty <manu@cbio.mskcc.org>
}
\references{
}
\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
