\name{dinuc.alphabet}
\alias{dinuc.alphabet}
\title{Convert to dinucleotide alphabet}
\usage{
  dinuc.alphabet(seqs)
}
\arguments{
  \item{seqs}{DNAStringSet object of sequences to be
  converted to dinucleotide}
}
\value{
  BStringSet object of dinucleotides
}
\description{
  Convert to dinucleotide alphabet
}

